part1

+++ part1/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: kayoshid <kayoshid@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/04/25 15:04:57 by kayoshid          #+#    #+#             */
-/*   Updated: 2026/04/29 18:46:55 by kayoshid         ###   ########.fr       */
+/*   Updated: 2026/04/30 02:47:26 by kayoshid         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,19 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	len;
+	size_t	len;
 
 	len = 0;
 	while (str[len] != '\0')
 		len++;
 	return (len);
 }
+
+// int	main()
+// {
+// 	char	str[] = "Hello World!";
+// 	size_t	n;
+// 	n = ft_strlen(str);
+// 	printf("%zu", n);
+// 	return (0);
+// }
